get up-to-date

+++ src/main/java/teammates/logic/core/CoursesLogic.java
@@ -1,7 +1,6 @@
 package teammates.logic.core;
 
 import java.time.Instant;
-import java.time.ZoneId;
 import java.util.ArrayList;
 import java.util.HashMap;
 import java.util.HashSet;
@@ -10,6 +9,7 @@ import java.util.Map;
 import java.util.Set;
 import java.util.stream.Collectors;
 
+import teammates.common.datatransfer.AttributesDeletionQuery;
 import teammates.common.datatransfer.CourseDetailsBundle;
 import teammates.common.datatransfer.CourseSummaryBundle;
 import teammates.common.datatransfer.FeedbackSessionDetailsBundle;
@@ -57,6 +57,9 @@ public final class CoursesLogic {
 
     private static final AccountsLogic accountsLogic = AccountsLogic.inst();
     private static final FeedbackSessionsLogic feedbackSessionsLogic = FeedbackSessionsLogic.inst();
+    private static final FeedbackQuestionsLogic fqLogic = FeedbackQuestionsLogic.inst();
+    private static final FeedbackResponsesLogic frLogic = FeedbackResponsesLogic.inst();
+    private static final FeedbackResponseCommentsLogic frcLogic = FeedbackResponseCommentsLogic.inst();
     private static final InstructorsLogic instructorsLogic = InstructorsLogic.inst();
     private static final StudentsLogic studentsLogic = StudentsLogic.inst();
 
@@ -68,42 +71,51 @@ public final class CoursesLogic {
         return instance;
     }
 
-    public void createCourse(String courseId, String courseName, String courseTimeZone)
+    /**
+     * Creates a course.
+     *
+     * @return the created course
+     * @throws InvalidParametersException if the course is not valid
+     * @throws EntityAlreadyExistsException if the course already exists in the Datastore.
+     */
+    public CourseAttributes createCourse(CourseAttributes courseToCreate)
             throws InvalidParametersException, EntityAlreadyExistsException {
-
-        CourseAttributes courseToAdd = validateAndCreateCourseAttributes(courseId, courseName, courseTimeZone);
-        coursesDb.createEntity(courseToAdd);
+        return coursesDb.createEntity(courseToCreate);
     }
 
     /**
-     * Creates a Course object and an Instructor object for the Course.
+     * Creates a course and an associated instructor for the course.
+     *
+     * <br/>Preconditions: <br/>
+     * * {@code instructorGoogleId} already has an account and instructor privileges.
      */
-    public void createCourseAndInstructor(String instructorGoogleId, String courseId, String courseName,
-                                          String courseTimeZone)
+    public void createCourseAndInstructor(String instructorGoogleId, CourseAttributes courseToCreate)
             throws InvalidParametersException, EntityAlreadyExistsException {
 
         AccountAttributes courseCreator = accountsLogic.getAccount(instructorGoogleId);
-        Assumption.assertNotNull("Trying to create a course for a non-existent instructor :" + instructorGoogleId,
-                                 courseCreator);
-        Assumption.assertTrue("Trying to create a course for a person who doesn't have instructor privileges :"
-                                  + instructorGoogleId,
-                              courseCreator.isInstructor);
+        Assumption.assertNotNull(
+                "Trying to create a course for a non-existent instructor :" + instructorGoogleId, courseCreator);
+        Assumption.assertTrue(
+                "Trying to create a course for a person who doesn't have instructor privileges :" + instructorGoogleId,
+                courseCreator.isInstructor());
 
-        createCourse(courseId, courseName, courseTimeZone);
+        CourseAttributes createdCourse = createCourse(courseToCreate);
 
         /* Create the initial instructor for the course */
         InstructorPrivileges privileges = new InstructorPrivileges(
                 Const.InstructorPermissionRoleNames.INSTRUCTOR_PERMISSION_ROLE_COOWNER);
         InstructorAttributes instructor = InstructorAttributes
-                .builder(instructorGoogleId, courseId, courseCreator.name, courseCreator.email)
+                .builder(createdCourse.getId(), courseCreator.getEmail())
+                .withName(courseCreator.getName())
+                .withGoogleId(instructorGoogleId)
                 .withPrivileges(privileges)
                 .build();
 
         try {
             instructorsLogic.createInstructor(instructor);
         } catch (EntityAlreadyExistsException | InvalidParametersException e) {
-            //roll back the transaction
-            coursesDb.deleteCourse(courseId);
+            // roll back the transaction
+            coursesDb.deleteCourse(createdCourse.getId());
             String errorMessage = "Unexpected exception while trying to create instructor for a new course "
                                   + System.lineSeparator() + instructor.toString() + System.lineSeparator()
                                   + TeammatesException.toStringWithStackTrace(e);
@@ -151,6 +163,11 @@ public final class CoursesLogic {
     public List<CourseDetailsBundle> getCourseDetailsListForStudent(String googleId)
                 throws EntityDoesNotExistException {
 
+        List<StudentAttributes> studentDataList = studentsLogic.getStudentsForGoogleId(googleId);
+        if (studentDataList.isEmpty()) {
+            throw new EntityDoesNotExistException("Student with Google ID " + googleId + " does not exist");
+        }
+
         List<CourseAttributes> courseList = getCoursesForStudentAccount(googleId);
         CourseAttributes.sortById(courseList);
         List<CourseDetailsBundle> courseDetailsList = new ArrayList<>();
@@ -194,32 +211,13 @@ public final class CoursesLogic {
     }
 
     /**
-     * Returns a list of section names for the course with ID courseId.
-     */
-    public List<String> getSectionsNameForCourse(String courseId) throws EntityDoesNotExistException {
-        return getSectionsNameForCourse(courseId, false);
-    }
-
-    /**
-     * Returns a list of section names for the specified course.
-     */
-    public List<String> getSectionsNameForCourse(CourseAttributes course) throws EntityDoesNotExistException {
-        Assumption.assertNotNull("Course is null", course);
-        return getSectionsNameForCourse(course.getId(), true);
-    }
-
-    /**
-     * Returns a list of section names for a course with or without a need to
-     * check if the course is existent.
+     * Returns a list of section names for the course with valid ID courseId.
      *
      * @param courseId Course ID of the course
-     * @param isCourseVerified Determine whether it is necessary to check if the course exists
      */
-    private List<String> getSectionsNameForCourse(String courseId, boolean isCourseVerified)
-            throws EntityDoesNotExistException {
-        if (!isCourseVerified) {
-            verifyCourseIsPresent(courseId);
-        }
+    public List<String> getSectionsNameForCourse(String courseId) throws EntityDoesNotExistException {
+        verifyCourseIsPresent(courseId);
+
         List<StudentAttributes> studentDataList = studentsLogic.getStudentsForCourse(courseId);
 
         Set<String> sectionNameSet = new HashSet<>();
@@ -411,7 +409,7 @@ public final class CoursesLogic {
         Assumption.assertNotNull("Supplied parameter was null", cd);
 
         CourseDetailsBundle cdd = new CourseDetailsBundle(cd);
-        cdd.sections = (ArrayList<SectionDetailsBundle>) getSectionsForCourse(cd, cdd);
+        cdd.sections = getSectionsForCourse(cd, cdd);
 
         return cdd;
     }
@@ -468,12 +466,8 @@ public final class CoursesLogic {
      *
      * @param googleId The Google ID of the student
      */
-    public List<CourseAttributes> getCoursesForStudentAccount(String googleId) throws EntityDoesNotExistException {
+    public List<CourseAttributes> getCoursesForStudentAccount(String googleId) {
         List<StudentAttributes> studentDataList = studentsLogic.getStudentsForGoogleId(googleId);
-
-        if (studentDataList.isEmpty()) {
-            throw new EntityDoesNotExistException("Student with Google ID " + googleId + " does not exist");
-        }
 
         List<String> courseIds = studentDataList.stream()
                 .filter(student -> !getCourse(student.course).isCourseDeleted())
@@ -594,7 +588,7 @@ public final class CoursesLogic {
     public Map<String, CourseDetailsBundle> getCourseSummariesForInstructor(
             List<InstructorAttributes> instructorAttributesList) {
 
-        HashMap<String, CourseDetailsBundle> courseSummaryList = new HashMap<>();
+        Map<String, CourseDetailsBundle> courseSummaryList = new HashMap<>();
         List<String> courseIdList = new ArrayList<>();
 
         for (InstructorAttributes instructor : instructorAttributesList) {
@@ -634,82 +628,71 @@ public final class CoursesLogic {
     }
 
     /**
-     * Updates the course details.
-     * @param courseId Id of the course to update
-     * @param courseName new name of the course
-     * @param courseTimeZone new time zone of the course
+     * Updates a course by {@link CourseAttributes.UpdateOptions}.
+     *
+     * <p>If the {@code timezone} of the course is changed, cascade the change to its corresponding feedback sessions.
+     *
+     * @return updated course
+     * @throws InvalidParametersException if attributes to update are not valid
+     * @throws EntityDoesNotExistException if the course cannot be found
      */
-    public void updateCourse(String courseId, String courseName, String courseTimeZone)
+    public CourseAttributes updateCourseCascade(CourseAttributes.UpdateOptions updateOptions)
             throws InvalidParametersException, EntityDoesNotExistException {
-        CourseAttributes newCourse = validateAndCreateCourseAttributes(courseId, courseName, courseTimeZone);
-        CourseAttributes oldCourse = coursesDb.getCourse(newCourse.getId());
+        CourseAttributes oldCourse = coursesDb.getCourse(updateOptions.getCourseId());
+        CourseAttributes updatedCourse = coursesDb.updateCourse(updateOptions);
 
-        if (oldCourse == null) {
-            throw new EntityDoesNotExistException("Trying to update a course that does not exist.");
+        if (!updatedCourse.getTimeZone().equals(oldCourse.getTimeZone())) {
+            feedbackSessionsLogic
+                    .updateFeedbackSessionsTimeZoneForCourse(updatedCourse.getId(), updatedCourse.getTimeZone());
         }
 
-        coursesDb.updateCourse(newCourse);
-        if (!newCourse.getTimeZone().equals(oldCourse.getTimeZone())) {
-            feedbackSessionsLogic.updateFeedbackSessionsTimeZoneForCourse(newCourse.getId(), newCourse.getTimeZone());
-        }
+        return updatedCourse;
     }
 
     /**
-     * Permanently deletes a course in Recycle Bin by its given corresponding ID.
-     * This will also cascade the data in other databases which are related to this course.
+     * Deletes a course cascade its students, instructors, sessions, responses and comments.
+     *
+     * <p>Fails silently if no such course.
      */
     public void deleteCourseCascade(String courseId) {
-        studentsLogic.deleteStudentsForCourse(courseId);
-        instructorsLogic.deleteInstructorsForCourse(courseId);
-        feedbackSessionsLogic.deleteFeedbackSessionsForCourseCascade(courseId);
+        if (getCourse(courseId) == null) {
+            return;
+        }
+
+        AttributesDeletionQuery query = AttributesDeletionQuery.builder()
+                .withCourseId(courseId)
+                .build();
+        frcLogic.deleteFeedbackResponseComments(query);
+        frLogic.deleteFeedbackResponses(query);
+        fqLogic.deleteFeedbackQuestions(query);
+        feedbackSessionsLogic.deleteFeedbackSessions(query);
+        studentsLogic.deleteStudents(query);
+        instructorsLogic.deleteInstructors(query);
+
         coursesDb.deleteCourse(courseId);
     }
 
     /**
-     * Permanently deletes all courses in Recycle Bin.
-     * This will also cascade the data in other databases which are related to these courses.
-     */
-    public void deleteAllCoursesCascade(List<InstructorAttributes> instructorList) {
-        Assumption.assertNotNull("Supplied parameter was null", instructorList);
-
-        List<String> softDeletedCourseIdList = instructorList.stream()
-                .filter(instructor -> coursesDb.getCourse(instructor.courseId).isCourseDeleted())
-                .map(InstructorAttributes::getCourseId)
-                .collect(Collectors.toList());
-
-        for (String courseId : softDeletedCourseIdList) {
-            deleteCourseCascade(courseId);
-        }
-    }
-
-    /**
      * Moves a course to Recycle Bin by its given corresponding ID.
-     * @return Soft-deletion time of the course.
+     * @return the time when the course is moved to the recycle bin
      */
-    public Instant moveCourseToRecycleBin(String courseId)
-            throws InvalidParametersException, EntityDoesNotExistException {
-        CourseAttributes course = coursesDb.getCourse(courseId);
-        course.setDeletedAt();
-        coursesDb.updateCourse(course);
+    public Instant moveCourseToRecycleBin(String courseId) throws EntityDoesNotExistException {
 
-        return course.deletedAt;
+        return coursesDb.softDeleteCourse(courseId);
     }
 
     /**
      * Restores a course from Recycle Bin by its given corresponding ID.
      */
-    public void restoreCourseFromRecycleBin(String courseId)
-            throws InvalidParametersException, EntityDoesNotExistException {
-        CourseAttributes course = coursesDb.getCourse(courseId);
-        course.resetDeletedAt();
-        coursesDb.updateCourse(course);
+    public void restoreCourseFromRecycleBin(String courseId) throws EntityDoesNotExistException {
+        coursesDb.restoreDeletedCourse(courseId);
     }
 
     /**
      * Restores all courses from Recycle Bin.
      */
     public void restoreAllCoursesFromRecycleBin(List<InstructorAttributes> instructorList)
-            throws InvalidParametersException, EntityDoesNotExistException {
+            throws EntityDoesNotExistException {
         Assumption.assertNotNull("Supplied parameter was null", instructorList);
 
         List<String> softDeletedCourseIdList = instructorList.stream()
@@ -725,7 +708,7 @@ public final class CoursesLogic {
     private Map<String, CourseSummaryBundle> getCourseSummaryWithoutStatsForInstructor(
             List<InstructorAttributes> instructorAttributesList) {
 
-        HashMap<String, CourseSummaryBundle> courseSummaryList = new HashMap<>();
+        Map<String, CourseSummaryBundle> courseSummaryList = new HashMap<>();
 
         List<String> courseIdList = instructorAttributesList.stream()
                 .filter(instructor -> !coursesDb.getCourse(instructor.courseId).isCourseDeleted())
@@ -818,36 +801,5 @@ public final class CoursesLogic {
             }
         }
         return archivedCourseIds;
-    }
-
-    /**
-     * Checks that {@code courseTimeZone} is valid and then returns a {@link CourseAttributes}.
-     * Field validation is usually done in {@link CoursesDb} by calling {@link CourseAttributes#getInvalidityInfo()}.
-     * However, a {@link CourseAttributes} cannot be created with an invalid time zone string.
-     * Hence, validation of this field is carried out here.
-     *
-     * @throws InvalidParametersException containing error messages for all fields if {@code courseTimeZone} is invalid
-     */
-    private CourseAttributes validateAndCreateCourseAttributes(
-            String courseId, String courseName, String courseTimeZone) throws InvalidParametersException {
-
-        // Imitate `CourseAttributes.getInvalidityInfo`
-        FieldValidator validator = new FieldValidator();
-        String timeZoneErrorMessage = validator.getInvalidityInfoForTimeZone(courseTimeZone);
-        if (!timeZoneErrorMessage.isEmpty()) {
-            // Leave validation of other fields to `CourseAttributes.getInvalidityInfo`
-            CourseAttributes dummyCourse = CourseAttributes
-                    .builder(courseId, courseName, Const.DEFAULT_TIME_ZONE)
-                    .build();
-            List<String> errors = dummyCourse.getInvalidityInfo();
-            errors.add(timeZoneErrorMessage);
-            // Imitate exception throwing in `CoursesDb`
-            throw new InvalidParametersException(errors);
-        }
-
-        // If time zone field is valid, leave validation  of other fields to `CoursesDb` like usual
-        return CourseAttributes
-                .builder(courseId, courseName, ZoneId.of(courseTimeZone))
-                .build();
     }
 }
